requirements.txt
Django==4.0.6
djangorestframework
djangorestframework-simplejwt
requests
django-redis
python-dotenv